requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2==0.2.0
google-auth-oauthlib>=0.4.6
notion-client>=2.0.0
python-dotenv>=0.19.0
python-telegram-bot>=13.7 